fmt: put shard entry in proper module
